requirements.txt
python-telegram-bot>=20.7
anthropic>=0.40.0
tavily-python>=0.5.0
supabase>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
